test: extract ScreenshotTestSupport (screenshot mocks) from TestSupport (#552, #434)
Second slice of splitting the 1,048-line TestSupport.swift "junk drawer"
into domain files, one cohesive mock group per PR, names stable, green
between. Slice 434a (RecordingTestSupport) shipped in #551.

Moves the two screenshot mocks — MockScreenshotCaptureService
(ScreenshotCapturing) and the @MainActor MockScreenshotSelectionService
(ScreenshotSelecting) — byte-identical from TestSupport.swift lines
18-85 into a new ScreenshotTestSupport.swift with the standard test
imports. The @MainActor annotation on the selection service is preserved
verbatim.

MockIssueExtractionService and MockExportService (both actors, sitting
at TestSupport lines 87+) explicitly do NOT move — they belong to a
separate slice (#557, 434g) and stay in TestSupport.swift for now. Every
other mock is untouched.

Pure test-support move: no product code, no helper renames, no behavior
change. The full BugNarratorTests suite stays green at 662 tests
(unchanged count from the current main baseline); docs-sync validation
passes.

Pre-build spec reviewed by codex in a fresh session — 5 findings folded
in, including the critical mis-scope catch: MockScreenshotSelectionService
is 33 lines (53-85), not the ~315 my initial coarse survey suggested.
The interstitial range 87-368 is MockIssueExtractionService +
MockExportService, correctly left in place and filed separately.
